Adapter is done

+++ app/src/main/java/com/example/simpleweatherapp/data/model/cities_model/CityModelItem.kt
@@ -3,7 +3,7 @@ package com.example.simpleweatherapp.data.model.cities_model
 data class CityModelItem(
     val coord: Coord,
     val country: String,
-    val id: Double,
+    val id: Long,
     val name: String,
     val state: String
 )
